+++ CHANGELOG.md
@@ -8,6 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ...
 
+## Changed
+ 
+- Database patching user interface presents clearer information about what version upgrade is occuring and the patches that will be applied.
+
+## Fixed
+
+- Database patching user interface no longer suggests restarting if the patching process has failed
+
 ## [3.2.0] - 2019-09-16
 
 ## Added